requirements.txt
Django==1.8.6
django-filter==0.11.0
djangorestframework==3.3.1
freeze==1.0.9
gunicorn==19.3.0
Markdown==2.6.4
psycopg2==2.6.1
six==1.10.0
wheel==0.24.0
